fixed a bug in new code to  keep track of total weight

+++ android-tegra-3.1/kernel/sched.c
@@ -433,7 +433,7 @@ struct wrr_rq {
 	 * It's NULL if nothing is running */
 	struct sched_wrr_entity *curr;
 	/* the total weight of the tasks in this run queue*/
-	unsigned long total_weight;
+	unsigned int total_weight;
 
 };
